Make the setup code for Jar-In-Jar reusable
